+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-exclude: ^src/wheel/vendored
+exclude: ^src/wheel/_vendored
 
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
