+++ pyproject.toml
@@ -32,7 +32,7 @@ Changelog = "https://wheel.readthedocs.io/en/stable/news.html"
 Source = "https://github.com/pypa/wheel"
 
 [project.scripts]
-wheel = "wheel.cli:main"
+wheel = "wheel._cli:main"
 
 [project.entry-points."distutils.commands"]
 bdist_wheel = "wheel.bdist_wheel:bdist_wheel"
@@ -70,7 +70,7 @@ exclude = [
 
 [tool.black]
 extend-exclude = '''
-^/src/wheel/vendored/
+^/src/wheel/_vendored/
 '''
 
 [tool.pytest.ini_options]
@@ -86,7 +86,7 @@ testpaths = ["test"]
 
 [tool.coverage.run]
 source = ["wheel"]
-omit = ["*/vendored/*"]
+omit = ["*/_vendored/*"]
 
 [tool.coverage.report]
 show_missing = true
