+++ src/wheel/_vendored/packaging/_elffile.py
