+++ src/wheel/_vendored/packaging/_structures.py
