+++ tests/_cli/eggnames.txt
