+++ tests/test_macosx_libfile.py
@@ -5,8 +5,8 @@ import struct
 import sysconfig
 
 import pytest
-from wheel.bdist_wheel import get_platform
-from wheel.macosx_libfile import extract_macosx_min_system_version
+from wheel._bdist_wheel import get_platform
+from wheel._macosx_libfile import extract_macosx_min_system_version
 
 
 def test_read_from_dylib():
